pyproject.toml
[tool.poetry]
name = "hydra-mlflow-trainer"
version = "0.1.0"
description = ""
authors = ["José Coelho <16445494+jcoelho93@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
mlflow = "^3.8.0"
hydra-core = "^1.3.2"
torch = "^2.9.1"
transformers = "^4.57.3"
datasets = "^4.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
